// frontend/src/interfaces/IRegisterUser/IRegisterUser.ts
export interface IRegisterUser {
	name: string;
	email: string;
	cpf: string;
	phone: string;
	birthday: string;
	description: string;
	advertiser: boolean;
	password: string;
	confirmPassword: string;
	cep: string;
	state: string;
	city: string;
	street: string;
	number: string;
	complement: string;
}